Move deployment description into server
