+++ host-controller/src/main/java/org/jboss/as/domain/controller/operations/deployment/DeploymentUploadURLHandler.java
@@ -21,10 +21,10 @@ package org.jboss.as.domain.controller.operations.deployment;
 import org.jboss.as.controller.OperationContext;
 import org.jboss.as.controller.OperationFailedException;
 import org.jboss.as.controller.descriptions.DescriptionProvider;
-import org.jboss.as.controller.descriptions.common.DeploymentDescription;
 import org.jboss.as.controller.operations.validation.ParametersValidator;
 import org.jboss.as.controller.operations.validation.StringLengthValidator;
 import org.jboss.as.repository.ContentRepository;
+import org.jboss.as.server.controller.descriptions.DeploymentDescription;
 import org.jboss.dmr.ModelNode;
 
 import java.io.IOException;
