+++ server/src/main/java/org/jboss/as/server/deployment/DeploymentUploadStreamAttachmentHandler.java
@@ -21,11 +21,11 @@ package org.jboss.as.server.deployment;
 import org.jboss.as.controller.OperationContext;
 import org.jboss.as.controller.OperationFailedException;
 import org.jboss.as.controller.descriptions.DescriptionProvider;
-import org.jboss.as.controller.descriptions.common.DeploymentDescription;
 import org.jboss.as.controller.operations.validation.IntRangeValidator;
 import org.jboss.as.controller.operations.validation.ParametersValidator;
 import org.jboss.as.repository.ContentRepository;
 import org.jboss.as.server.ServerMessages;
+import org.jboss.as.server.controller.descriptions.DeploymentDescription;
 import org.jboss.dmr.ModelNode;
 
 import java.io.IOException;
