+++ server/src/main/resources/org/jboss/as/server/controller/descriptions/LocalDescriptions.properties
@@ -116,3 +116,34 @@ shutdown=Shuts down the server via a call to System.exit(0)
 shutdown.restart=If true, once shutdown the server will be restarted again
 dump-services=Dumps all services running in the container, including their status and dependency information.
 restart-required=Puts the server into a restart-required mode.
+
+# Deployments
+deployment=A deployment represents anything that can be deployed (e.g. an application such as EJB-JAR, WAR, EAR, any kind of standard archive such as RAR or JBoss-specific deployment) into a server.
+deployment.name=Unique identifier of the deployment. Must be unique across all deployments.
+deployment.runtime-name=Name by which the deployment should be known within a server's runtime. This would be equivalent to the file name of a deployment file, and would form the basis for such things as default Java Enterprise Edition application and module names. This would typically be the same as 'name', but in some cases users may wish to have two deployments with the same 'runtime-name' (e.g. two versions of "foo.war") both available in the deployment content repository, in which case the deployments would need to have distinct 'name' values but would have the same 'runtime-name'.
+deployment.content=List of pieces of content that comprise the deployment.
+deployment.hash=The hash of managed deployment content that has been uploaded to the domain's or standalone server's deployment content repository.
+deployment.inputstream=The index into the operation's attached input streams of the input stream that contains deployment content that should be uploaded to the domain's or standalone server's deployment content repository.
+deployment.enabled=Boolean indicating whether the deployment content is currently deployed in the runtime (or should be deployed in the runtime the next time the server starts.)
+deployment.persistent=Boolean indicating whether the existence of the deployment should be recorded in the persistent server configuration. Only relevant to a standalone mode server. Default is 'true'. A deployment managed by a deployment scanner would have this set to 'false' to ensure the deployment is only deployed at server start if the scanner again detects the deployment.
+deployment.upload-bytes=Indicates that the deployment content in the included byte array should be added to the deployment content repository. Note that this operation does not indicate the content should be deployed into the runtime.
+deployment.bytes=Byte array containing the deployment content that should uploaded to the domain's or standalone server's deployment content repository..
+deployment.upload-url=Indicates that the deployment content available at the included URL should be added to the deployment content repository. Note that this operation does not indicate the content should be deployed into the runtime.
+deployment.url=The URL at which the deployment content is available for upload to the domain's or standalone server's deployment content repository.. Note that the URL must be accessible from the target of the operation (i.e. the Domain Controller or standalone server).
+deployment.upload-stream=Indicates that the deployment content available at the included input stream index should be added to the deployment content repository. Note that this operation does not indicate the content should be deployed into the runtime.
+deployment.path=Path (relative or absolute) to unmanaged content that is part of the deployment.
+deployment.relative-to=Name of a system path to which the value of the 'path' is relative. If not set, the 'path' is considered to be absolute.
+deployment.archive=Flag indicating whether unmanaged content is a zip archive (true) or exploded (false).
+deployment.add=Adds previously uploaded content to the list of content available for use. Does not actually deploy the content unless the 'enabled' parameter is 'true'.
+deployment.deploy=Deploy the specified deployment content into the runtime, optionally replacing existing content.
+deployment.deploy.name=The name of the new content.
+deployment.deploy.to-replace=The name of the content that is to be replaced.
+deployment.replace=Replace existing content in the runtime with new content. The new content must have been previously uploaded to the deployment content repository.
+deployment.replace.name=The name of the new content.
+deployment.replace.to-replace=The name of the content that is to be replaced.
+deployment.full-replace=Add previously uploaded deployment content to the list of content available for use, replace existing content of the same name in the runtime, and remove the replaced content from the list of content available for use. This is equivalent to an 'add', 'undeploy', 'deploy', 'remove' sequence where the new content has the same name as the content being replaced. Precisely one of 'runtime-name', 'hash', 'input-stream-index', 'bytes' or 'url' must be specified.
+deployment.undeploy=Undeploy content from the runtime. The content remains in the list of content available for use.
+deployment.redeploy=Undeploy existing content from the runtime and deploy it again.
+deployment.status=The current runtime status of a deployment. Possible status modes are OK, FAILED, and STOPPED. FAILED indicates a dependency is missing or a service could not start. STOPPED indicates that the deployment was not enabled or was manually stopped.
+deployment.subsystem=Runtime resources created when the deployment is deployed, organized by the subsystem responsible for the runtime resource.
+deployment.subdeployment=Runtime resources associated with a child deployment packaged inside another deployment; for example a war packaged inside an ear.
